+++ apps/docs/examples/BasicButton/ExampleBasicButtonDefault.js
@@ -1,7 +1,0 @@
-import React from "react";
-import {BasicButton} from "@productplan/atlas";
-
-/** Default style */
-export default function ExampleBasicButtonDefault() {
-  return <BasicButton label="Howdy" htmlId="default_button" htmlForm="" htmlType="button" type="default" />
-}