+++ apps/docs/examples/BasicButton/ExampleBasicButtonGhost.js
@@ -1,7 +1,0 @@
-import React from "react";
-import {BasicButton} from "@productplan/atlas";
-
-/** Ghost style */
-export default function ExampleBasicButtonGhost() {
-  return <BasicButton label="Howdy again" htmlId="ghost_button" htmlForm="" htmlType="button" type="ghost" />
-}