+++ apps/docs/examples/BasicButton/ExampleBasicButtonWithLeadingIcon.js
@@ -1,0 +1,16 @@
+import { BasicButton } from "@productplan/atlas";
+
+/** Button with a leading icon */
+export default function ExampleBasicButtonWithLeadingIcon() {
+  return (
+    <>
+      <BasicButton label="I have an icon!" htmlType="button" type="default" leadingIcon={<i className="fas fa-check" />} />
+      <BasicButton label="I have an icon!" htmlType="button" type="secondary" leadingIcon={<i className="fas fa-check" />} />
+      <BasicButton label="I have an icon!" htmlType="button" type="tertiary" leadingIcon={<i className="fas fa-check" />} />
+      <BasicButton label="I have an icon!" htmlType="button" type="destructive" leadingIcon={<i className="fas fa-check" />} />
+      <BasicButton label="I have an icon!" htmlType="button" type="ghost" leadingIcon={<i className="fas fa-check" />} />
+      <BasicButton label="I have an icon!" htmlType="button" type="default" leadingIcon={<i className="fas fa-check" />} size="small" />
+      <BasicButton label="I have an icon!" htmlType="button" type="default" leadingIcon={<i className="fas fa-check" />} size="mini" />
+    </>
+  );
+}