+++ apps/docs/examples/BasicButton/ExampleBasicButtonWithTrailingIcon.js
@@ -1,7 +1,16 @@
-import React from "react";
-import {BasicButton} from "@productplan/atlas";
+import { BasicButton } from "@productplan/atlas";
 
 /** Button with a trailing icon */
 export default function ExampleBasicButtonWithTrailingIcon() {
-  return <BasicButton label="I have an icon!" htmlForm="" htmlId="button_trailing_icon" htmlType="button" type="default" trailingIcon={<i className="fas fa-check" />} />
+  return (
+    <>
+      <BasicButton label="I have an icon!" type="default" trailingIcon={<i className="fas fa-check" />} />
+      <BasicButton label="I have an icon!" type="secondary" trailingIcon={<i className="fas fa-check" />} />
+      <BasicButton label="I have an icon!" type="tertiary" trailingIcon={<i className="fas fa-check" />} />
+      <BasicButton label="I have an icon!" type="destructive" trailingIcon={<i className="fas fa-check" />} />
+      <BasicButton label="I have an icon!" type="ghost" trailingIcon={<i className="fas fa-check" />} />
+      <BasicButton label="I have an icon!" type="default" trailingIcon={<i className="fas fa-check" />} size="small" />
+      <BasicButton label="I have an icon!" type="default" trailingIcon={<i className="fas fa-check" />} size="mini" />
+    </>
+  );
 }