+++ packages/atlas-web-components/src/components/BasicButton.jsx
@@ -1,29 +1,49 @@
-import register from 'preact-custom-element';
+import register from "preact-custom-element";
 import styles from "./BasicButton.scss";
 import "redefine-custom-elements";
 
 const BasicButton = ({
+  align = "center",
+  disabled = undefined,
+  htmlId = undefined,
   label = "",
+  leadingIcon = null,
+  size = "default",
   theme = "default",
+  trailingIcon = null,
   type = "button",
-  htmlId,
-  leadingIcon,
-  trailingIcon,
 }) => {
   return (
     <>
       <style>{styles}</style>
       <button
-        className={`BasicButton BasicButton--${theme}`}
+        className={`BasicButton BasicButton--${theme} BasicButton--size-${size} BasicButton--align-${align}`}
         type={type}
         htmlId={htmlId}
+        disabled={disabled}
+        size={size}
       >
-        {leadingIcon && <span className="leadingIcon">{leadingIcon}</span>}
+        <span className="leadingIcon"><slot name="leadingIcon"></slot></span>
         {label}
-        {trailingIcon && <span className="trailingIcon">{trailingIcon}</span>}
+        <span className="trailingIcon"><slot name="trailingIcon"></slot></span>
       </button>
     </>
   );
 };
 
-register(BasicButton, 'atlas-basic-button', [], { shadow: true });
+register(
+  BasicButton,
+  "atlas-basic-button",
+  [
+    "align",
+    "disabled",
+    "htmlId",
+    "label",
+    "leadingIcon",
+    "size",
+    "theme",
+    "trailingIcon",
+    "type",
+  ],
+  { shadow: true }
+);