+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @productplan/atlas-docs
 
+## 0.2.0
+
+### Minor Changes
+
+- eeb2322: add tertiary, destructive button and disable state
+
+### Patch Changes
+
+- Updated dependencies [eeb2322]
+  - @productplan/atlas@0.2.0
+  - @productplan/atlas-web-components@0.2.0
+
 ## 0.1.0
 
 ### Minor Changes